-added phow fast extraction with dll and staff
Former-commit-id: ac3c95d75b1695d24e6f825c35079e20312da3c7 [formerly 71c59a0a4e3273c3bfd0c61a8cef745f2a04884a]
Former-commit-id: 33dc4ef8ed8ce596c17284e4d5a7e8551e89446e
Former-commit-id: 1828384dd08237926d2927f714de15733f45a3a4 [formerly 4fa14a2b4a5d89b65e28d8652976ba60d21b4180]
Former-commit-id: fc902d6aa87c84ae5df44703d32bab345c9f858f
Former-commit-id: 244680e57ad028be655a8156686e0dff70bfbee4 [formerly 3166f2be56c83a22a9619c05a698301256515bed]
Former-commit-id: 39a470d5351b5ca70a0ea2694c6d4efc103eaeb9

+++ AutomaticImageClassificationTests/FeaturesTest.cs
@@ -44,7 +44,7 @@ namespace AutomaticImageClassificationTests
 
             IKdTree tree = new VlFeatKdTree();
             tree.CreateTree(model.Means);
-            phow = new VlFeatPhow(tree, model.Means[0].Length, 256, 256);
+            phow = new VlFeatPhow(tree, model.Means[0].Length, 256, 256,true);
 
             foreach (var image in sampleImgs)
             {
@@ -414,29 +414,41 @@ namespace AutomaticImageClassificationTests
             var testFile = @"Data\Features\" + feature + "_Tree" + tree + "_Cluster" + cluster + "_" + clusterNum + "_test.txt";
             var clustersFile = @"Data\Palettes\" + feature + "_" + clusterNum + "_clusters.txt";
 
-            #region Clustering
 
-            var sampleImgs = Files.GetFilesFrom(trainPath, sampleImages);
-            var clusterFeaturesPerImage = maxNumberClusterFeatures / sampleImgs.Length;
-
-            var clusters = new List<double[]>();
-            foreach (var image in sampleImgs)
+            List<double[]> finalClusters;
+            if (File.Exists(clustersFile))
             {
-                clusters.AddRange(feature.ExtractDescriptors(image).OrderBy(x => Guid.NewGuid()).Take(clusterFeaturesPerImage));
+                #region read Clusters from File
+
+                finalClusters = Files.ReadFileToListArrayList<double>(clustersFile);
+
+                #endregion
             }
-            sampleImgs = null;
-            ClusterModel model = cluster.CreateClusters(clusters, clusterNum);
-            clusters.Clear();
+            else
+            {
+                #region Clustering
 
-            tree.CreateTree(model.Means);
+                var sampleImgs = Files.GetFilesFrom(trainPath, sampleImages);
+                var clusterFeaturesPerImage = maxNumberClusterFeatures / sampleImgs.Length;
 
-            List<double[]> finalClusters = model.Means;
-            Files.WriteFile(clustersFile, finalClusters);
+                var clusters = new List<double[]>();
+                foreach (var image in sampleImgs)
+                {
+                    clusters.AddRange(feature.ExtractDescriptors(image).OrderBy(x => Guid.NewGuid()).Take(clusterFeaturesPerImage));
+                }
+                sampleImgs = null;
+                ClusterModel model = cluster.CreateClusters(clusters, clusterNum);
+                clusters.Clear();
+                finalClusters = model.Means;
 
-            #endregion
 
-            feature = new VlFeatPhow(tree, clusterNum);
-            //finalClusters.Clear();
+                Files.WriteFile(clustersFile, finalClusters);
+
+                #endregion
+            }
+
+            //tree.CreateTree(finalClusters);
+            feature = new VlFeatPhow(finalClusters,true);
 
             #region features extraction
